Remote server reconfigure
