+++ Capstone Project/Project Code/client/style.css
@@ -1,10 +1,11 @@
-body{
+body {
   min-width: 98vw;
   max-width: 98vw;
   min-height: 98vh;
   max-height: 98vh;
   overflow: hidden;
 }
+
 *:focus {
   outline: none;
 }
@@ -80,24 +81,18 @@ body{
   background-color: white;
   padding: 10px;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-  
   border-style: solid;
   border-width: 2px;
   background-color: whitesmoke;
-  
-  
 }
+
 .header input[type=text] {
   background-color: white;
-
   border: none;
   outline: none;
   font-size: 17px;
   width: 90vw;
-  
 }
-
-
 
 /* The grid container */
 .grid-container {
@@ -108,7 +103,6 @@ body{
     'left main-content main-content main-content main-content main-content main-content main-content main-content main-content';
   grid-column-gap: 10px; 
   height: auto;
-
 } 
 
 .left {
@@ -125,19 +119,17 @@ body{
   background-color: whitesmoke;
   min-width: 12vw;
   max-width: 12vw;
-  }
+}
 
 /* Style the main-content class */
 .main-content {
   grid-area: main-content;
-  
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
   border-style: solid;
   border-width: 2px;
   background-color: whitesmoke;
   text-align: left;
   display: flex;
-  
   height: 88vh;
   max-height: 88vh;
   max-width: 86vw;
@@ -148,16 +140,16 @@ body{
 
 /* Responsive layout - makes the three columns stack on top of each other instead of next to each other */
 @media (max-width: 600px) {
-    .grid-container  {
-      grid-template-areas: 
-    'header header header header header header header header header header' 
-    'left main-content main-content main-content main-content main-content main-content main-content main-content main-content ';
-    }
+  .grid-container  {
+    grid-template-areas: 
+  'header header header header header header header header header header' 
+  'left main-content main-content main-content main-content main-content main-content main-content main-content main-content ';
+  }
 
-    #col3 {
-      visibility: hidden;
-    }
-  }  
+  #col3 {
+    visibility: hidden;
+  }
+}  
 
 .accordion {
   background-color: whitesmoke;
@@ -173,7 +165,6 @@ body{
   font-size:large;
   font-weight: bold;
   border-bottom-style:ridge;
-  
 }
 
 .accordion:after {
@@ -204,8 +195,6 @@ body{
   border-right:black;
   overflow-x: scroll;
   overflow-y: hidden;
-  
-  
 }
 
 .column {
